+++ vscode-react/src/components/atoms/index.ts
@@ -1,0 +1,3 @@
+export { Button } from "@react-components/atoms/button.component";
+export { Logo } from "@react-components/atoms/logo.component";
+export { Overlay } from "@react-components/atoms/overlay.component";
