+++ vscode-react/src/components/atoms/table/index.ts
@@ -1,1 +1,3 @@
-export { TableCell } from "@react-components/atoms/table/cell.component";
+export { Cell } from "@react-components/atoms/table/cell.component";
+export { HeaderCell } from "@react-components/atoms/table/headerCell.component";
+export { TableMessage } from "@react-components/atoms/table/tableMessage.component";
