+++ vscode-react/src/components/deployments/atoms/index.ts
@@ -1,0 +1,1 @@
+export { DeploymentStateLabel } from "@react-components/deployments/atoms/stateLabel.component";
