+++ vscode-react/src/components/deployments/index.ts
@@ -1,0 +1,3 @@
+export { DeploymentStateLabel } from "@react-components/deployments/atoms";
+export { DeploymentTableHeader } from "@react-components/deployments/molecules";
+export { ExecutePopper, DeploymentsTableBody } from "@react-components/deployments/organisms";
