+++ vscode-react/src/components/deployments/molecules/deploymentTableHeader.component.tsx
@@ -1,18 +1,19 @@
 import React from "react";
 import { translate } from "@i18n";
-import { TableHeader, TableHeaderCell } from "@react-components/Table";
+import { HeaderCell } from "@react-components/atoms/table";
+import { TableHeader } from "@react-components/molecules/table/tableHeader.component";
 
 export const DeploymentTableHeader: React.FC = () => {
 	return (
 		<TableHeader classes="sticky top-12 z-30">
-			<TableHeaderCell>{translate().t("reactApp.deployments.time")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.deployments.status")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.sessions.statuses.stopped")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.sessions.statuses.running")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.sessions.statuses.error")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.sessions.statuses.completed")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.deployments.buildId")}</TableHeaderCell>
-			<TableHeaderCell>{translate().t("reactApp.deployments.actions")}</TableHeaderCell>
+			<HeaderCell>{translate().t("reactApp.deployments.time")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.deployments.status")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.sessions.statuses.stopped")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.sessions.statuses.running")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.sessions.statuses.error")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.sessions.statuses.completed")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.deployments.buildId")}</HeaderCell>
+			<HeaderCell>{translate().t("reactApp.deployments.actions")}</HeaderCell>
 		</TableHeader>
 	);
 };
