+++ vscode-react/src/components/deployments/molecules/index.ts
@@ -1,0 +1,1 @@
+export { DeploymentTableHeader } from "@react-components/deployments/molecules/deploymentTableHeader.component";
