+++ vscode-react/src/components/deployments/organisms/index.ts
@@ -1,0 +1,2 @@
+export { ExecutePopper } from "@react-components/deployments/organisms/executePopper.component";
+export { DeploymentsTableBody } from "@react-components/deployments/organisms/deploymentsTableBody.component";
