+++ vscode-react/src/components/deployments/pages/deploymentsSection.component.tsx
@@ -9,7 +9,7 @@ import { useIncomingMessageHandler } from "@react-hooks";
 import { sendMessage } from "@react-utilities";
 import { Deployment } from "@type/models";
 
-export const Deployments = ({ height }: { height: string | number }) => {
+export const DeploymentsSection = ({ height }: { height: string | number }) => {
 	const [deploymentsSection, setDeploymentsSection] = useState<DeploymentSectionViewModel>();
 	const [totalDeployments, setTotalDeployments] = useState<number>();
 	const [deployments, setDeployments] = useState<Deployment[]>();
