+++ vscode-react/src/components/deployments/pages/index.ts
@@ -1,0 +1,1 @@
+export { Deployments } from "@react-components/deployments/pages/deploymentsSection.component";
