+++ vscode-react/src/components/molecules/index.ts
@@ -1,0 +1,4 @@
+export { DeletePopper } from "@react-components/molecules/deletePopper.component";
+export { Modal } from "@react-components/molecules/modal.component";
+export { Popper } from "@react-components/molecules/popper.component";
+export { TableRow, TableHeader } from "@react-components/molecules/table";
