+++ vscode-react/src/components/molecules/table/index.ts
@@ -1,0 +1,2 @@
+export { TableHeader } from "@react-components/molecules/table/tableHeader.component";
+export { TableRow } from "@react-components/molecules/table/tableRow.component";
