+++ vscode-react/src/components/organisms/index.ts
@@ -1,0 +1,3 @@
+export { Header } from "@react-components/organisms/header.component";
+export { MonacoEditorModal } from "@react-components/organisms/monacoEditorModal.component";
+export { Table } from "@react-components/organisms/table.component";
