+++ vscode-react/src/components/organisms/monacoEditorModal.component.tsx
