+++ vscode-react/src/components/project/organisms/index.ts
@@ -1,0 +1,1 @@
+export { ProjectSettingsPopper } from "@react-components/project/organisms/projectSettingsPopper.component";
