+++ vscode-react/src/components/sessions/atoms/index.ts
@@ -1,0 +1,1 @@
+export { SessionStateLabel } from "@react-components/sessions/atoms/sessionStateLabel.component";
