+++ vscode-react/src/components/sessions/molecules/index.ts
@@ -1,0 +1,1 @@
+export { SessionActions } from "@react-components/sessions/molecules/sessionActions.component";
