+++ vscode-react/src/components/sessions/organisms/index.ts
@@ -1,0 +1,2 @@
+export { SessionsTableBody } from "@react-components/sessions/organisms/sessionsTableBody.component";
+export { SessionsTableRow } from "@react-components/sessions/organisms/sessionsTableRow.component";
