+++ vscode-react/src/components/sessions/organisms/sessionsTableBody.component.tsx
