+++ vscode-react/src/components/sessions/organisms/sessionsTableRow.component.tsx
