+++ vscode-react/src/components/sessions/pages/index.ts
@@ -1,0 +1,1 @@
+export { SessionsSection } from "@react-components/sessions/pages/sessionsSection.component";
